+++ modules/names/utils/getFullNames.ts
@@ -1,14 +1,20 @@
-import firstNamesList from "../data/firstNamesList";
-import lastNamesList from "../data/lastNamesList";
-import {getRandomArrayItem} from "../../../utils/arrays";
+import { faker } from '@faker-js/faker';
 
-const getFullNames = (qty: number) : string[] => { 
+const getFullNames = (qty: number, sex: string = "both") : string[] => { 
     const namesList = [];
 
     for (let index = 0; index < qty; index++) {
-        let randomFirstName = getRandomArrayItem(firstNamesList);
-        let randomLastName = getRandomArrayItem(lastNamesList);
-        namesList.push(`${randomFirstName} ${randomLastName}`);
+        switch (sex) {
+            case "female":
+                namesList.push(faker.name.fullName({sex : "female"}))
+                break;
+            case "male":
+                namesList.push(faker.name.fullName({sex : "male"}))
+                break;
+            default:
+                namesList.push(faker.name.fullName())
+                break;
+        }
     }
 
     return namesList;