+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to `ageddesi/Mocked-AP` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.3] - 2022-08-21
+### Added
+- Added faker.js as a data source and update 'Names' endpoint to use faker.js
+- Added new Names end points for middleName and all endpoint now have an additional 'sex' url added
+
 ## [v0.0.2] - 2022-08-19
 ### Added
 - New 'Countries' Endpoint: